HermiteOptimalControl.jl Documentation

HermiteOptimalControl.jl Documentation

Workflow

Functions

Schrodinger Problem Definition

Schrodinger Problem Examples

Forward Evolution

Gradient Evaluation

Bsplines

Workflow

Right now the user can set up Schrodinger problems, and compute the gradients of them with the control vector and target gate of their choosing. In a complete package, there would also be an optimization procedure which uses this gradient calculation.

The basic workflow is:

Set up a SchrodingerProblem, either yourself using the constructor SchrodingerProb, or by using one of the example problems provided. Schrodinger Problem Examples

Choose a control vector and target.

Compute a gradient using one of the methods provided. Gradient Evaluation

Functions

Schrodinger Problem Definition

SchrodingerProb

Schrodinger Problem Examples

Missing docstring.

Missing docstring for rabi_osc. Check Documenter's build log for details.

Missing docstring.

Missing docstring for gargamel_prob. Check Documenter's build log for details.

Missing docstring.

Missing docstring for bspline_prob. Check Documenter's build log for details.

Forward Evolution

Functions for evolving the state vector in a problem forward in time according to Schrodinger's equation, with or without forcing.

QuantumGateDesign.eval_forward — Function

kwargs should be propogated all the way down to the vector version of eval_forward.

Evolve a vector SchrodingerProblem forward in time. Return the history of the state vector (u/v) in a 3-index array, where the first index of corresponds to the vector component, the second index corresponds to the derivative to be taken, and the third index corresponds to the timestep number.

Missing docstring.

Missing docstring for eval_forward_forced. Check Documenter's build log for details.

Gradient Evaluation

QuantumGateDesign.discrete_adjoint — Function

Arbitrary order version, should make one with target being abstract vector as well, so I can do state transfer problems.

Missing docstring.

Missing docstring for eval_grad_forced. Check Documenter's build log for details.

Missing docstring.

Missing docstring for eval_grad_finite_difference. Check Documenter's build log for details.

QuantumGateDesign.infidelity — Function

Calculates the infidelity for the given state vector 'ψ' and target state 'target.'

Returns: Infidelity

Calculates the infidelity for the given matrix of state vectors 'Q' and matrix of target states 'target.'

Returns: Infidelity

Bsplines

Missing docstring.

Missing docstring for bcparams. Check Documenter's build log for details.

Missing docstring.

Missing docstring for bcarrier2. Check Documenter's build log for details.

Missing docstring.

Missing docstring for bcarrier2_dt. Check Documenter's build log for details.

Missing docstring.

Missing docstring for gradbcarrier2!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for gradbcarrier2_dt!. Check Documenter's build log for details.