HermiteOptimalControl.jl Documentation

HermiteOptimalControl.jl Documentation

Workflow

Functions

Schrodinger Problem Definition

Schrodinger Problem Examples

Forward Evolution

Gradient Evaluation

Bsplines

Workflow

Right now the user can set up Schrodinger problems, and compute the gradients of them with the control vector and target gate of their choosing. In a complete package, there would also be an optimization procedure which uses this gradient calculation.

The basic workflow is:

Set up a SchrodingerProblem, either yourself using the constructor SchrodingerProb, or by using one of the example problems provided. Schrodinger Problem Examples

Choose a control vector and target.

Compute a gradient using one of the methods provided. Gradient Evaluation

Functions

Schrodinger Problem Definition

SchrodingerProb

Schrodinger Problem Examples

Missing docstring.

Missing docstring for rabi_osc. Check Documenter's build log for details.

Missing docstring.

Missing docstring for gargamel_prob. Check Documenter's build log for details.

Missing docstring.

Missing docstring for bspline_prob. Check Documenter's build log for details.

Forward Evolution

Functions for evolving the state vector in a problem forward in time according to Schrodinger's equation, with or without forcing.

HermiteOptimalControl.eval_forward — Function

Missing docstring.

Missing docstring for eval_forward_forced. Check Documenter's build log for details.

Gradient Evaluation

HermiteOptimalControl.discrete_adjoint — Function

Evaluates gradient of the provided Schrodinger problem with the given target gate and control parameter(s) pcof using the discrete adjoint method.

Returns: gradient

HermiteOptimalControl.eval_grad_forced — Function

Evaluates gradient of the provided Schrodinger problem with the given target gate and control parameter(s) pcof using the "forward differentiation" method, which evolves a differentiated Schrodinger equation, using the state vector in the evolution of the original Schrodinger equation as a forcing term.

Returns: gradient

HermiteOptimalControl.eval_grad_finite_difference — Function

Vector control version.

Evaluates gradient of the provided Schrodinger problem with the given target gate and control parameter(s) pcof using a finite difference method, where a step size of dpcof is used when perturbing the components of the control vector pcof.

Returns: gradient

HermiteOptimalControl.infidelity — Function

Calculates the infidelity for the given state vector 'ψ' and target state 'target.'

Returns: Infidelity

Calculates the infidelity for the given matrix of state vectors 'Q' and matrix of target states 'target.'

Returns: Infidelity

Bsplines

HermiteOptimalControl.bcparams — Type

bcpar = bcparams(T, D1, Ncoupled, Nunc, omega, pcof)

General constructor of struct bcparams for setting up B-splines with carrier waves.

bcpar = bcparams(T, D1, omega, pcof)

Simplified constructor for the case when there are no uncoupled controls and Ncoupled = size(omega,1).

Arguments

T:: Float64: Duration of spline function

D1:: Int64: Number of basis functions in each segment

Ncoupled::Int64: Number of coupled controls in the simulation

Nunc::Int64: Number of uncoupled controls in the simulation

omega::Array{Float64,2}: Carrier wave frequencies

pcof:: Array{Float64, 1}: Coefficient vector. Must have D1*Nseg elements

First dimensions of the omega array:

Without uncoupled controls, Nunc=0 and size(omega,1) = Ncoupled.

With uncoupled controls, Nunc > 0 and size(omega,1) = Ncoupled + Nunc.

Second dimension of the omega array:

size(omega, 2) = Nfreq

Ordering of the pcof array:

First consider the case without uncoupled control functions, Nunc = 0: Then the pcof array then has 2*Ncoupled*Nfreq*D1 elements. Each ctrl ∈ [1,Ncoupled] and freq ∈ [1,Nfreq] corresponds to D1 elements in the pcof vector. For the case Ncoupled = 2 and Nfreq = 2, the elements are ordered according to

ctrl	freq	α_1	α_2

1	1	1:D1	D1+1:2 D1

1	2	2 D1+1: 3 D1	3 D1+1:4 D1

2	1	4 D1+1: 5 D1	5 D1+1:6 D1

2	2	6 D1+1: 7 D1	7 D1+1: 8D1

If there are uncoupled controls, Nunc > 0, the pcof array should have (2*Ncoupled + Nunc)*Nfreq*D1 elements. The last Nunc*Nfreq*D1 elements correspond to the uncoupled control functions and are ordered in a corresponding way.

External links

Spline Wavelet on Wikipedia.

HermiteOptimalControl.bcarrier2 — Function

f = bcarrier2(t, params, func)

Evaluate a B-spline function with carrier waves. See also the bcparams constructor.

Arguments

t::Float64: Evaluate spline at parameter t ∈ [0, param.T]

param::params: Parameters for the spline

func::Int64: Spline function index ∈ [0, param.Nseg-1]

f = bcarrier2(t, params, func)

Evaluate a B-spline function with carrier waves. See also the bcparams constructor.

Arguments

t::Float64: Evaluate spline at parameter t ∈ [0, param.T]

param::params: Parameters for the spline

func::Int64: Spline function index ∈ [0, param.Nseg-1]

HermiteOptimalControl.bcarrier2_dt — Function

f = bcarrier2_dt(t, params, func)

Evaluate a B-spline function with carrier waves. See also the bcparams constructor.

Arguments

t::Float64: Evaluate spline at parameter t ∈ [0, param.T]

param::params: Parameters for the spline

func::Int64: Spline function index ∈ [0, param.Nseg-1]

SPENCER: To get cross derivative (d2/dtda, just remove the pcofs in the below, since the controls are linear in them)

f = bcarrier2_dt(t, params, func)

Evaluate a B-spline function with carrier waves. See also the bcparams constructor.

Arguments

t::Float64: Evaluate spline at parameter t ∈ [0, param.T]

param::params: Parameters for the spline

func::Int64: Spline function index ∈ [0, param.Nseg-1]

SPENCER: To get cross derivative (d2/dtda, just remove the pcofs in the below, since the controls are linear in them)

HermiteOptimalControl.gradbcarrier2! — Function

gradbcarrier2!(t, bcpar, func, g) -> g

Evaluate the gradient of a control function with respect to all coefficient.

NOTE: the index of the control functions is 0-based. For a set of coupled controls, mod(func,2)=0 corresponds to ∇ pj(t) and mod(func,2) = 1 corresponds to ∇ qj(t), where j = div(func,2).

Arguments

t::Float64: Evaluate spline at parameter t ∈ [0, param.T]

bcpar::bcparams: Parameters for the spline

func::Int64: Control function index ∈ [0, param.Nseg-1]

g::Array{Float64,1}: Preallocated array to store calculated gradient

Missing docstring.

Missing docstring for gradbcarrier2_dt!. Check Documenter's build log for details.